+++ starviewer/src/inputoutput/qcreatedicomdir.h
@@ -143,8 +143,8 @@ private slots:
     void deviceChanged( int value );
 
 private:
-    double m_dicomdirSize;
-    double m_DiskSpace;
+    quint64 m_dicomdirSize;
+    quint64 m_DiskSpace;
     QMenu m_contextMenu;///<Menu contextual
 
     /// Dispositius on es pot gravar el DICOMDIR
